move Default driver into its own module
